requirements.txt
streamlit>=1.28.0
pymongo>=4.6.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
dnspython>=2.4.2